+++ lib/features/login/logic/cubit/login_cubit.dart
@@ -1,5 +1,4 @@
 import 'package:doctor_appointment/features/login/data/models/login_request_body.dart';
-import 'package:doctor_appointment/features/login/data/models/login_response.dart';
 import 'package:doctor_appointment/features/login/data/repos/login_repo.dart';
 import 'package:doctor_appointment/features/login/logic/cubit/login_state.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
@@ -10,8 +9,8 @@ class LoginCubit extends Cubit<LoginState> {
   void emitLoginStates(LoginRequestBody loginRequestBody) async {
     emit(const LoginState.loading());
     final response = await _loginRepo.login(loginRequestBody);
-    response.when(success: (LoginResponse) {
-      emit(LoginState.success(loginRequestBody));
+    response.when(success: (loginResponse) {
+      emit(LoginState.success(loginResponse));
     }, failure: (error) {
       emit(LoginState.error(error));
     });
