+++ lib/features/on_boarding/ui/widgets/doc_image_on_boarding.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 
-import '../../../core/theming/text_styles.dart';
+import '../../../../core/theming/text_styles.dart';
 
 class DoctorImage extends StatelessWidget {
   const DoctorImage({super.key});
